moved timetravel command to commands module
