+++ src/commands/commands.js
@@ -1,3 +1,5 @@
+import Election from "../election";
+
 /**
  * @summary changes user access level (can only de-elevate)
  * @param {import("../index").BotConfig} config bot config
@@ -30,4 +32,38 @@ export const setAccessCommand = (config, user, content) => {
     changeMap[level]?.();
 
     return `Changed access level of ${uid} to ${level}`;
+};
+
+/**
+ * @summary changes internal bot clock to a given day
+ * @param {Election} election current election instance
+ * @param {string} content incoming message content
+ * @returns {string}
+ */
+export const timetravelCommand = (election, content) => {
+    const [, yyyy, MM, dd, today] = /(?:(\d{4})-(\d{2})-(\d{2}))|(today)/.exec(content) || [];
+
+    if (!today && (!yyyy || !MM || !dd)) return "Sorry, Doc! Invalid coordinates";
+
+    const destination = today ? new Date() : new Date(+yyyy, +MM - 1, +dd);
+
+    const phase = Election.getPhase(election, destination);
+
+    election.phase = phase;
+
+    const intl = new Intl.DateTimeFormat("en-US", {
+        year: "numeric",
+        month: "short",
+        day: "2-digit",
+        hour12: true,
+        hour: "2-digit",
+        minute: "2-digit"
+    });
+
+    const arrived = intl
+        .format(destination)
+        .replace(/, /g, " ")
+        .replace(/ (?:AM|PM)$/, "");
+
+    return `Arrived at ${arrived}, today's phase: ${phase || "no phase"}`;
 };
